+++ src/main/java/com/example/planeo_back/web/DTO/expense/ExpenseByTagDTO.java
@@ -2,5 +2,7 @@ package com.example.planeo_back.web.DTO.expense;
 
 import com.example.planeo_back.domain.enums.Tag;
 
-public record ExpenseByTagDTO(Tag tag, Double total) {
+import java.math.BigDecimal;
+
+public record ExpenseByTagDTO(Tag tag, BigDecimal total) {
 }
